Add service update provider (with Montreal STM Bus implementation) & code cleaning.

+++ src/org/mtransit/android/commons/provider/StatusFilter.java
@@ -50,7 +50,7 @@ public abstract class StatusFilter implements MTLog.Loggable {
 	}
 
 	public boolean hasCacheValidityInMs() {
-		return cacheValidityInMs != null && cacheValidityInMs > 0;
+		return cacheValidityInMs != null && cacheValidityInMs.longValue() > 0;
 	}
 
 	public void setCacheValidityInMs(Long cacheValidityInMs) {
